requirements.txt
numpy
pandas
scipy
librosa
soundfile
noisereduce
webrtcvad-wheels
pyannote.audio
torch
scikit-learn
spidev
matplotlib